pyproject.toml
[project]
name = "smde"
version = "0.1.0"
description = "Smart Maritime Document Extractor backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13,<2.0",
    "asyncpg>=0.29,<1.0",
    "fastapi>=0.111,<1.0",
    "google-genai>=1.0,<2.0",    "httpx>=0.27,<1.0",
    "pydantic-settings>=2.3,<3.0",
    "pydantic>=2.7,<3.0",
    "python-multipart>=0.0.9,<1.0",
    "sqlalchemy>=2.0,<3.0",
    "uvicorn>=0.30,<1.0",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=0.23,<1.0",
    "pytest>=8.0,<9.0",
    "ruff>=0.5,<1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-p no:cacheprovider"

[tool.ruff]
line-length = 100
target-version = "py313"

